+++ packages/graphql-codegen-typescript-operations/src/visitor.ts
@@ -9,9 +9,9 @@ import {
   LoadedFragment,
   normalizeAvoidOptionals,
   ParsedDocumentsConfig,
-  PreResolveTypesProcessor,
   SelectionSetProcessorConfig,
   SelectionSetToObject,
+  PreResolveTypesProcessor as CodegenPreResolveTypesProcessor,
   wrapTypeWithModifiers,
 } from "@graphql-codegen/visitor-plugin-common";
 import autoBind from "auto-bind";
@@ -31,6 +31,7 @@ import {
 import { TypeScriptDocumentsPluginConfig } from "./config";
 import { TypeScriptOperationVariablesToObject } from "./ts-operation-variables-to-object";
 import { TypeScriptSelectionSetProcessor } from "./ts-selection-set-processor";
+import { PreResolveTypesProcessor } from "./ts-pre-resolve-types-processor";
 
 export interface TypeScriptDocumentsParsedConfig extends ParsedDocumentsConfig {
   avoidOptionals: AvoidOptionalsConfig;
@@ -45,6 +46,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
   TypeScriptDocumentsParsedConfig
 > {
   private usedTypes: Set<string>;
+  private usedEnums: Set<string>;
   private isExactUsed = false;
 
   constructor(
@@ -71,6 +73,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
     autoBind(this);
 
     this.usedTypes = new Set();
+    this.usedEnums = new Set();
     const wrapOptional = (type: string) => {
       const prefix =
         !this.config.inlineCommonTypes && this.config.namespacedImportName
@@ -91,6 +94,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
     ): string => {
       const optional =
         !this.config.avoidOptionals.field && !!type && !isNonNullType(type);
+
       return (
         (this.config.immutableTypes ? `readonly ${name}` : name) +
         (optional ? "?" : "")
@@ -100,7 +104,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
     const processorConfig: SelectionSetProcessorConfig = {
       namespacedImportName: this.config.namespacedImportName,
       convertName: this.convertName.bind(this),
-      enumPrefix: this.config.enumPrefix,
+      enumPrefix: false,
       scalars: this.scalars,
       formatNamedField,
       wrapTypeWithModifiers(baseType, type) {
@@ -111,9 +115,12 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
       },
       avoidOptionals: this.config.avoidOptionals,
     };
+
     const processor = new (
       config.preResolveTypes
-        ? PreResolveTypesProcessor
+        ? config.inlineCommonTypes
+          ? PreResolveTypesProcessor
+          : CodegenPreResolveTypesProcessor
         : TypeScriptSelectionSetProcessor
     )(processorConfig);
     this.setSelectionSetHandler(
@@ -142,6 +149,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
         this.config.enumValues,
         true,
         this.config.inlineCommonTypes,
+        this.usedEnums,
       ),
     );
     this._declarationBlockConfig = {
@@ -149,6 +157,15 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
     };
   }
 
+  private getExportNames(): string[] {
+    const exportNames: string[] = Array.from(this.usedTypes);
+
+    if (!this.config.inlineCommonTypes) {
+      exportNames.push(...this.usedEnums);
+    }
+
+    return exportNames;
+  }
   public getImports(): Array<string> {
     const imports = !this.config.globalNamespace
       ? this.config.fragmentImports
@@ -157,9 +174,11 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
           )
           .concat(
             this.config.baseTypesPath && this.usedTypes.size
-              ? `export ${this.config.isTypeOnly ? "type " : ""}{${Array.from(
-                  this.usedTypes,
-                ).join(",")}} from "${this.config.baseTypesPath}"`
+              ? `export ${
+                  this.config.isTypeOnly ? "type " : ""
+                }{${this.getExportNames().join(",")}} from "${
+                  this.config.baseTypesPath
+                }"`
               : "",
           )
       : [];
@@ -168,6 +187,18 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
       imports.push(`type Exact<T extends { [key: string]: unknown }> = {
         [K in keyof T]: T[K];
       }`);
+    }
+
+    if (this.config.inlineCommonTypes && this.usedEnums.size) {
+      for (const usedEnum of this.usedEnums.values()) {
+        const enumType = this.schema.getType(usedEnum) as GraphQLEnumType;
+        imports.push(
+          `export type ${usedEnum} = ${enumType
+            .getValues()
+            .map((value) => `"${value.name}"`)
+            .join(" | ")}`,
+        );
+      }
     }
 
     return imports;
@@ -218,7 +249,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
       } else if (
         this.schema.getType(fieldType.name) instanceof GraphQLEnumType
       ) {
-        this.usedTypes.add(fieldType.name);
+        this.usedEnums.add(fieldType.name);
       }
     }
   }
@@ -243,7 +274,7 @@ export class TypeScriptDocumentsVisitor extends BaseDocumentsVisitor<
         this.getAllEntitiesRelatedToInput(astNode);
       }
     } else if (this.schema.getType(convertedName) instanceof GraphQLEnumType) {
-      this.usedTypes.add(convertedName);
+      this.usedEnums.add(convertedName);
     }
 
     return super.convertName(node, options);